Use data base

+++ src/CashFlow.DataProvider.EFCore/Providers/AssetProvider.cs
@@ -19,8 +19,7 @@ namespace CashFlow.DataProvider.EFCore.Providers
 
         public async Task<IReadOnlyCollection<AssetDto>> GetAllAssetForUserAsync(string userId)
         {
-            //TODO: Тест выборка
-            return await _context.Assets.Where(x => x.UserId == userId || x.UserId == null || x.UserId == string.Empty)
+            return await _context.Assets.Where(x => x.UserId == userId)
                 .OrderBy(x => x.OrderNumber).AsNoTracking()
                 .Select(x => new AssetDto { Id = x.Id, Name = x.Name, OrderNumber = x.OrderNumber, Price = x.Price, Quantity = x.Quantity })
                 .ToArrayAsync();
@@ -46,11 +45,10 @@ namespace CashFlow.DataProvider.EFCore.Providers
         public async Task<AssetDto> CreateAssetForUserAsync(AssetDto asset, string userId)
         {
             //TODO: бизенс логику убрать в menager, сделать в одной транзакции
-            //TODO: Тест выборка
             var orderNumber = _context.Incomes
-                .Where(x => x.UserId == userId || x.UserId == null || x.UserId == string.Empty)
+                .Where(x => x.UserId == userId)
                 .OrderByDescending(x => x.OrderNumber).FirstOrDefault()?.OrderNumber + 1 ?? 1;
-            var newIncome = new Income { Name = asset.Name, OrderNumber = orderNumber, Price = 0 };
+            var newIncome = new Income { Name = asset.Name, OrderNumber = orderNumber, Price = 0 , UserId = userId };
             await _context.Incomes.AddAsync(newIncome);
 
             var newAsset = new Asset
@@ -79,7 +77,8 @@ namespace CashFlow.DataProvider.EFCore.Providers
             //TODO: бизенс логику убрать в menager, сделать в одной транзакции
             Asset original = await _context.Assets.FindAsync(id);
             Income income = await _context.Incomes.FindAsync(original.IncomeId);
-            _context.Incomes.Remove(income);
+            if(income != null)
+                _context.Incomes.Remove(income);
             _context.Assets.Remove(original);
             await _context.SaveChangesAsync();
         }
